requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
sentence-transformers
PyPDF2
pypdf
chromadb
groq
